test(jsonrpc): refactor simulator
