+++ tests/jsonrpc/simulator/namespaces/admin.go
