+++ tests/jsonrpc/simulator/namespaces/debug.go
