+++ tests/jsonrpc/simulator/namespaces/miner.go
