+++ tests/jsonrpc/simulator/namespaces/net.go
