+++ tests/jsonrpc/simulator/namespaces/personal.go
