+++ tests/jsonrpc/simulator/namespaces/websocket.go
